+++ src/lib.rs
@@ -4,13 +4,22 @@ use std::io::*;
 use std::fs::File;
 use std::fmt;
 
+/// Custom result with ChlogError
 type ChlogResult = std::result::Result<(), ChlogError>;
 
+/// Custom error type with errors specific to changelogger
 #[derive(Debug)]
 pub enum ChlogError {
+    /// Exercise is already present in the changelog file
     ExercisePresent,
+
+    /// Changelog file was not found
     FileNotFound,
+
+    /// Couldn't create a file
     FileWriteFailed(std::io::Error),
+
+    /// Couldn't save a file
     FileCreateFailed(std::io::Error),
 }
 
@@ -25,14 +34,15 @@ impl fmt::Display for ChlogError {
     }
 }
 
-// TODO: should return result with proper error handling
-pub fn add_exercise(language: &str, name: &str, source: &str, path: Option<&str>) -> ChlogResult {
+/// Function for adding an exercise to the changelog
+/// # Arguments:
+/// language: Programming language of the exercise
+/// name: Name of the exercise
+/// source: Source of the exercise (e.g. CodeWars)
+/// path: path to the changelog file
+pub fn add_exercise(language: &str, name: &str, source: &str, file_path: &str) -> ChlogResult {
+    // buffer for the file
     let mut buff = String::new();
-
-    let file_path = match path {
-        Some(s) => s,
-        None => "CHANGELOG.md",
-    };
 
     if let Ok(mut chlog) = File::open(file_path) {
         if let Ok(_) = chlog.read_to_string(&mut buff) {