+++ src/main.rs
@@ -1,7 +1,7 @@
-
-extern crate clap;
-use clap::{Arg, App};
 extern crate changelogger;
+extern crate clap;
+
+use clap::{Arg, App};
 
 fn main() {
     let matches = App::new("changelogger")
@@ -14,7 +14,14 @@ fn main() {
              .value_name("LANGUAGE")
              .value_name("TITLE")
              .value_name("SOURCE")
-             .help("Information about the exercise")
+             .help("Adds a new programming exercise")
+        .takes_value(true))
+        .arg(Arg::with_name("add_commit")
+             .short("c")
+             .long("add_commit")
+             .value_name("CATEGORY")
+             .value_name("DESCRIPTION")
+             .help("Adds a new development project commit")
         .takes_value(true))
         .arg(Arg::with_name("file")
              .short("f")
@@ -32,6 +39,14 @@ fn main() {
         };
 
         if let Err(e) = changelogger::add_exercise(data[0], data[1], data[2], path) {
+            println!("{}", e);
+        }
+    }
+
+    if let Some(commit) = matches.values_of("add_commit") {
+        let data: Vec<&str> = commit.collect();
+
+        if let Err(e) = changelogger::add_commit(data[0], data[1]) {
             println!("{}", e);
         }
     }